Adjusting The HTML & JS Code
A <div> was created in the HTML file which contains three (3) buttons.
The purpose of this is catching the click event while is bubbling up to <div>
element. Thus, setting one event listener instead of three on each button element.

Furthermore, "getHumanChoice" function was removed as at this version of the game
the player will click on the buttons we added previously instead of submitting his
choice via a prompt. Additionally, the "playRound" function and score variables were
put outside the "playGame" function, thus making it available globally.